+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+- Hide site menu on embed mode
+
+## 2.0.0 RC10
+- Date pickers years selectors
+- Date pickers now allows selection anywhere in the overall time range
+- Allow multiple polygons in report
+- Send layer name to report
+- Temporarily remove rendering of vessel track outside of inner time range
 - Don't resize viewport on toggle custom layer visibility
 - Fixed loader style on Safari
 - Pinned vessels: fixed a bug where typing a custom would be very slow when pinned track was visible